+++ fbw-a32nx/src/systems/extras-host/index.ts
@@ -13,6 +13,8 @@ import {
   PilotSeatManager,
   ExtrasSimVarPublisher,
   GsxSimVarPublisher,
+  GsxSyncA320,
+  MsfsFlightModelPublisher,
 } from '@flybywiresim/fbw-sdk';
 import { PushbuttonCheck } from 'extras-host/modules/pushbutton_check/PushbuttonCheck';
 import { FlightPlanAsoboSync } from 'extras-host/modules/flightplan_sync/FlightPlanAsoboSync';
@@ -20,7 +22,6 @@ import { KeyInterceptor } from './modules/key_interceptor/KeyInterceptor';
 import { VersionCheck } from './modules/version_check/VersionCheck';
 import { AircraftSync } from './modules/aircraft_sync/AircraftSync';
 import { LightSync } from 'extras-host/modules/light_sync/LightSync';
-import { GsxSync } from 'extras-host/modules/gsx_sync/GsxSync';
 
 /**
  * This is the main class for the extras-host instrument.
@@ -64,6 +65,8 @@ class ExtrasHost extends BaseInstrument {
 
   private readonly simVarPublisher: ExtrasSimVarPublisher;
 
+  private readonly msfsFlightModelPublisher: MsfsFlightModelPublisher;
+
   private readonly gsxVarPublisher: GsxSimVarPublisher;
 
   private readonly pushbuttonCheck: PushbuttonCheck;
@@ -82,7 +85,7 @@ class ExtrasHost extends BaseInstrument {
 
   private readonly lightSync: LightSync = new LightSync(this.bus);
 
-  private readonly gsxSync: GsxSync;
+  private readonly gsxSync = new GsxSyncA320(this.bus);
 
   /**
    * "mainmenu" = 0
@@ -95,9 +98,9 @@ class ExtrasHost extends BaseInstrument {
   constructor() {
     super();
 
-    this.bus = new EventBus();
     this.hEventPublisher = new HEventPublisher(this.bus);
     this.simVarPublisher = new ExtrasSimVarPublisher(this.bus);
+    this.msfsFlightModelPublisher = new MsfsFlightModelPublisher(this.bus);
     this.gsxVarPublisher = new GsxSimVarPublisher(this.bus);
 
     this.notificationManager = new NotificationManager(this.bus);
@@ -108,11 +111,14 @@ class ExtrasHost extends BaseInstrument {
 
     this.versionCheck = new VersionCheck(process.env.AIRCRAFT_PROJECT_PREFIX, this.bus);
     this.aircraftSync = new AircraftSync(process.env.AIRCRAFT_PROJECT_PREFIX, this.bus);
-    this.gsxSync = new GsxSync(this.bus);
+
+    this.backplane.addPublisher('GsxSimVarPublisher', this.gsxVarPublisher);
+    this.backplane.addPublisher('MsfsFlightModelPublisher', this.msfsFlightModelPublisher);
 
     this.backplane.addInstrument('Clock', this.clock);
     this.backplane.addInstrument('PilotSeatManager', this.pilotSeatManager);
     this.backplane.addInstrument('LightSync', this.lightSync);
+    this.backplane.addInstrument('GsxSync', this.gsxSync);
 
     console.log('A32NX_EXTRASHOST: Created');
   }
@@ -137,7 +143,6 @@ class ExtrasHost extends BaseInstrument {
     this.keyInterceptor.connectedCallback();
     this.flightPlanAsoboSync.connectedCallback();
     this.aircraftSync.connectedCallback();
-    this.gsxSync.connectedCallback();
 
     this.backplane.init();
   }
@@ -157,14 +162,12 @@ class ExtrasHost extends BaseInstrument {
         this.versionCheck.startPublish();
         this.keyInterceptor.startPublish();
         this.simVarPublisher.startPublish();
-        this.gsxVarPublisher.startPublish();
         this.flightPlanAsoboSync.init();
         this.aircraftSync.startPublish();
       }
       this.gameState = gs;
     } else {
       this.simVarPublisher.onUpdate();
-      this.gsxVarPublisher.onUpdate();
     }
 
     this.versionCheck.update();