+++ fbw-common/src/systems/shared/src/publishers/Msfs/index.ts
@@ -1,2 +1,3 @@
 export * from './MsfsAutopilotAssistancePublisher';
 export * from './MsfsRadioNavigationPublisher';
+export * from './MsfsFlightModelPublisher';