+++ fbw-common/src/systems/shared/src/publishers/index.ts
@@ -2,3 +2,4 @@ export * from './Msfs';
 export * from './OansBtv';
 export * from './FcuBusPublisher';
 export * from './Extras';
+export * from './GroundSupportPublisher';